// extension/views/constant.d.ts
export declare const DIST_WEBVIEW_PATH = "distWebview";
export declare const DIST_WEBVIEW_INDEX_HTML = "distWebview/index.html";
export declare const WEBVIEW_NAMES: {
    PreviewImages: string;
};
export declare const MESSAGE_CMD: {
    GET_ALL_IMGS: string;
    RENAME_FILE: string;
    DELETE_FILE: string;
    OPEN_IMAGE_DIRECTORY: string;
    GET_IMAGE_BASE64: string;
    GET_IMAGE_SIZE: string;
    SAVE_CONFIG: string;
    GET_CONFIG: string;
};
export declare const EXTENSION_NAME = "vscode-infra";
export declare const EXTENSION_COMMANDS: {
    OPEN_WEBVIEW_IMAGE_VIEWER: string;
};
export declare const BACKGROUND_COLOR_OPTIONS: string[];
export declare const DEFAULT_BACKGROUND_COLOR: string;
export declare const DEFAULT_IMAGE_SIZE = 100;